+++ osc/sine.c
@@ -15,7 +15,7 @@ void osc_sine_ticker(void * info) {
     me->phase += *(me->frequency) * PI * 2 / sample_rate;
     if ( me->phase > PI*2 ) me->phase -= PI*2;
 
-    me->now = cos(me->phase);
+    me->now = cosf(me->phase);
 }
 
 struct osc_sine_st * osc_sine_make(float phase, float *frequency) {
